feat game-53-java rudimentary port of basic to java
This version of game 53 in java was (predominantly) mechanically
translated to create one java method per basic-lang line.
This accomplishes a behaviorally similar game. The next commit
should refactor the source to something reasonable.

This commit includes a similarly generated graphviz dot-lang file
that represents the graph of basic lines. This aids in the visual
identification of flows of control (ex cycles represent loops, or
chains of subroutines).

-----BEGIN PGP SIGNED MESSAGE-----
Hash: SHA256

= Contributor license agreement - basic computer games
== Agreement
I, Nicholas Prado, using the username Nzen, Br-Nicholas-Prado,
or any other, affirm that I release all of my contributions to
the software project called "basic-computer-games" to the
public domain, or - where not recognized, under the terms of the
Unlicense license v1.0 or any later version, as accessible in
this project as the LICENSE, in the project root,
or at https://choosealicense.com/licenses/unlicense/.
- -- (signed) 2023-08-08
-----BEGIN PGP SIGNATURE-----